Fixed some coding style errors
