+++ src/routes/mod.rs
@@ -1,3 +1,4 @@
 pub mod download;
 pub mod index;
 pub mod pull;
+pub mod search;
